1. changed orm float to int 2. scheduled stock price change save to db

+++ orm/holdings.go
@@ -9,7 +9,7 @@ type Holdings struct {
 	UserID                   int64
 	StockID                  int64
 	Quantity                 int64
-	AverageCostPerShareCents float64
+	AverageCostPerShareCents int64
 	CreatedAt                time.Time
 	UpdatedAt                time.Time
 }
