+++ util/currency_util.go
@@ -7,10 +7,10 @@ import (
 
 const InitialInvestmentCents = 10000000
 
-func ConvertCentsToDollars(cents float64) float64 {
+func ConvertCentsToDollars(cents int64) float64 {
 
 	dollars := cents / 100.0
-	rounded := fmt.Sprintf("%.2f", dollars)
+	rounded := fmt.Sprintf("%d", dollars)
 	parsedDollar, err := strconv.ParseFloat(rounded, 64)
 	if err != nil {
 		return 0
